requirements.txt
python-dotenv==1.0.1
langchain==0.2.14
langchain-openai==0.1.21
langgraph==0.2.34
llama-index==0.11.23
